+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - The Publish All button in the editor reloads the entry list
 - `Number::fraction(num, denom)` constructor
 - Added negative exponent problems
+- Editor now has separate search field for each kind of entry
 
 ### Changed
 - Renamed `standard_powers` to `integer_exponents`